// src/app/models/orderdetail.ts
export class OrderDetail {

  entity_id: string;
  billing_address_id: string;
  shipping_address_id: string;
  coupon_code: string;
  subtotal: string;
  shipping_method_id: string;
  tax: string;
  total: string;
  storecredit: string;
  storecredit_amount: string;
  invoiceNumber: string;
  credit_memo: string;
  credit_memo_amt: string;
  discount_voucher: string;
  discount_voucher_amt: string;
  gift_voucher: string;
  gift_voucher_amt: string;
  total_payable: string;
  store_id: string;
  status: string;
  created_on: string;
  sales_order_id: string;
  total_paid: string;
  shipping_amount: string;
  increment_id: string;
  store_name: string;
  shipping_method: string;
}